Improve acceptance testing & fix bug when project id is empty (#7)
FEATURE:

- Add testing framework supported `scenario` and `step`, inspire by *ThoughtWorks, Guage*
- Add parameters encoding for array on general invoking

ENHANCEMENT:

- Add test scenario and improve the testing framework
- Improve continuous deployment workflow, integration on a private cluster
- Add unit testing cases to improving line coverage at core modules

BUGFIX:

- Fix unexpected error when project id is empty

+++ ucloud/services/udb/schemas/models.py
@@ -30,11 +30,17 @@ class UDBSlaveInstanceSetSchema(schema.ResponseSchema):
 
     fields = {
         "AdminUser": fields.Str(required=False, load_from="AdminUser"),
-        "BackupBeginTime": fields.Int(required=False, load_from="BackupBeginTime"),
-        "BackupBlacklist": fields.Str(required=False, load_from="BackupBlacklist"),
+        "BackupBeginTime": fields.Int(
+            required=False, load_from="BackupBeginTime"
+        ),
+        "BackupBlacklist": fields.Str(
+            required=False, load_from="BackupBlacklist"
+        ),
         "BackupCount": fields.Int(required=False, load_from="BackupCount"),
         "BackupDate": fields.Str(required=False, load_from="BackupDate"),
-        "BackupDuration": fields.Int(required=False, load_from="BackupDuration"),
+        "BackupDuration": fields.Int(
+            required=False, load_from="BackupDuration"
+        ),
         "ChargeType": fields.Str(required=False, load_from="ChargeType"),
         "ClusterRole": fields.Str(required=False, load_from="ClusterRole"),
         "CreateTime": fields.Int(required=False, load_from="CreateTime"),
@@ -46,7 +52,9 @@ class UDBSlaveInstanceSetSchema(schema.ResponseSchema):
         "ExpiredTime": fields.Int(required=False, load_from="ExpiredTime"),
         "InstanceMode": fields.Str(required=False, load_from="InstanceMode"),
         "InstanceType": fields.Str(required=False, load_from="InstanceType"),
-        "InstanceTypeId": fields.Int(required=False, load_from="InstanceTypeId"),
+        "InstanceTypeId": fields.Int(
+            required=False, load_from="InstanceTypeId"
+        ),
         "LogFileSize": fields.Float(required=False, load_from="LogFileSize"),
         "MemoryLimit": fields.Int(required=False, load_from="MemoryLimit"),
         "ModifyTime": fields.Int(required=False, load_from="ModifyTime"),
@@ -58,7 +66,9 @@ class UDBSlaveInstanceSetSchema(schema.ResponseSchema):
         "SrcDBId": fields.Str(required=False, load_from="SrcDBId"),
         "State": fields.Str(required=False, load_from="State"),
         "SubnetId": fields.Str(required=False, load_from="SubnetId"),
-        "SystemFileSize": fields.Float(required=False, load_from="SystemFileSize"),
+        "SystemFileSize": fields.Float(
+            required=False, load_from="SystemFileSize"
+        ),
         "Tag": fields.Str(required=False, load_from="Tag"),
         "UseSSD": fields.Bool(required=False, load_from="UseSSD"),
         "VPCId": fields.Str(required=False, load_from="VPCId"),
@@ -74,11 +84,17 @@ class UDBInstanceSetSchema(schema.ResponseSchema):
 
     fields = {
         "AdminUser": fields.Str(required=False, load_from="AdminUser"),
-        "BackupBeginTime": fields.Int(required=False, load_from="BackupBeginTime"),
-        "BackupBlacklist": fields.Str(required=False, load_from="BackupBlacklist"),
+        "BackupBeginTime": fields.Int(
+            required=False, load_from="BackupBeginTime"
+        ),
+        "BackupBlacklist": fields.Str(
+            required=False, load_from="BackupBlacklist"
+        ),
         "BackupCount": fields.Int(required=False, load_from="BackupCount"),
         "BackupDate": fields.Str(required=False, load_from="BackupDate"),
-        "BackupDuration": fields.Int(required=False, load_from="BackupDuration"),
+        "BackupDuration": fields.Int(
+            required=False, load_from="BackupDuration"
+        ),
         "BackupZone": fields.Str(required=False, load_from="BackupZone"),
         "ChargeType": fields.Str(required=False, load_from="ChargeType"),
         "CluserRole": fields.Str(required=False, load_from="CluserRole"),
@@ -92,7 +108,9 @@ class UDBInstanceSetSchema(schema.ResponseSchema):
         "ExpiredTime": fields.Int(required=False, load_from="ExpiredTime"),
         "InstanceMode": fields.Str(required=False, load_from="InstanceMode"),
         "InstanceType": fields.Str(required=False, load_from="InstanceType"),
-        "InstanceTypeId": fields.Int(required=False, load_from="InstanceTypeId"),
+        "InstanceTypeId": fields.Int(
+            required=False, load_from="InstanceTypeId"
+        ),
         "LogFileSize": fields.Float(required=False, load_from="LogFileSize"),
         "MemoryLimit": fields.Int(required=False, load_from="MemoryLimit"),
         "ModifyTime": fields.Int(required=False, load_from="ModifyTime"),
@@ -104,7 +122,9 @@ class UDBInstanceSetSchema(schema.ResponseSchema):
         "SrcDBId": fields.Str(required=False, load_from="SrcDBId"),
         "State": fields.Str(required=False, load_from="State"),
         "SubnetId": fields.Str(required=False, load_from="SubnetId"),
-        "SystemFileSize": fields.Float(required=False, load_from="SystemFileSize"),
+        "SystemFileSize": fields.Float(
+            required=False, load_from="SystemFileSize"
+        ),
         "Tag": fields.Str(required=False, load_from="Tag"),
         "UseSSD": fields.Bool(required=False, load_from="UseSSD"),
         "VPCId": fields.Str(required=False, load_from="VPCId"),
@@ -132,7 +152,7 @@ class UDBInstancePriceSetSchema(schema.ResponseSchema):
 
     fields = {
         "ChargeType": fields.Str(required=False, load_from="ChargeType"),
-        "Price": fields.Float(required=False, load_from="Price"),
+        "Price": fields.Int(required=False, load_from="Price"),
     }
